Add ProxyClientBase destroy_connection option
Give ProxyClientBase class the ability to delete the Connection object in its
destructor. This avoids the need for separate
[`ShutdownLoop`](https://github.com/ryanofsky/bitcoin/blob/pr/ipc/src/interfaces/capnp/ipc.cpp#L50-L65)
close hook in https://github.com/bitcoin/bitcoin/pull/10102.

This exposed a race condition which required adding a TaskSet member to
Connection objects to fix. With the previous ShutdownLoop approach, the client
capability handle would be freed in the event loop thread, then there would be
a switch back to the client thread, then there would be another switch back to
the event loop thread to free the connection, most likely after the event loop
thread had a chance to process new events. With the new approach, the client
capability handle and Connection object are destroyed back to back in the event
loop thread, causing the onDisconnect handler to fire when it previously
didn't, after the Connection object had already been destroyed, and resulting
in a double deletion of the Connection pointer. The race is fixed here by using
separate kj::TaskSet objects for each connection, so onDisconnect events are
now cancelled as Connection objects are destroyed.