Change EventLoop::m_task_set to not use boost
EventLoop::m_task_set was unnecessary defined as
boost::optional<kj::TaskSet> because it is always initialized.

It can, however, possibly be destroyed twice: in the EventLoop::loop()
method (conditional) and in the EventLoop destructor when it goes out of
scope (unconditional). So, use std::unique_ptr to handle this.