Disable clang suggest-override warnings for proxy clients
Warnings look like "warning: 'initError' overrides a member function but
is not marked 'override' [-Wsuggest-override]"

Warnings were originally reported by Sjors Provoost <sjors@sprovoost.nl>
and are tracked in
https://github.com/chaincodelabs/libmultiprocess/issues/61

The same warnings were previously disabled for GCC in
05f9817c369a5b74a9a802b0af063c103560a1cc
https://github.com/chaincodelabs/libmultiprocess/pull/40 and that change
is extended here to apply to clang. The methods triggering the warnings
are not virtual in all cases and should not have the override keyword.