Add ReadDestTemp function to make it easier to call ReadField with less boilerplate
Add RestDestTemp function to simplify ReadField callers that just want to read a
field into a new object. Instead of requiring ReadField callers to construct the
object in advance and pass with ReadDestValue, or to write custom emplace
function with ReadDestEmplace, they call RestDestTemp to make ReadField just
return a temporary object that can be moved from.

Add some test coverage for this. Also add missing documentation for
ReadDestEmplace and ReadDestValue classes.