mpgen: Avoid deprecated SchemaParser::parseDiskFile call
This avoids deprecation warnings, and is potentially a little safer since it
sandboxes the parsing and makes it an error to import paths outside the source
directory.

This change does have some downsides, though. It complicates include logic, and
also drops compatibility with capnproto versions before 0.7.0 that don't have
the new SchemaParser::parseFromDirectory method which was added in
https://github.com/capnproto/capnproto/commit/c1fe2b0339a1499c354eb85a2a78c63b2ce449f3
It also adds a hard dependency on the kj/filesystem.h library which was
previously optional.

Fixes #39