move: add mp/type-data.h
BuildField definition was moved from
https://github.com/bitcoin/bitcoin/blob/0a931a9787b196d7a620863cc143d9319ffd356d/src/ipc/capnp/common-types.h#L137-L162
and tweaked to be more restrictive so it only matches `LocalType`'s that are
convertible to spans and constructible from spans, not just types that are
convertible to spans. This allows adding a matching CustomReadField function
below, which is new.

Having the CustomReadField function fixes a serialization bug in the bitcoin
core mining IPC interface that was reported in
https://github.com/Sjors/bitcoin/issues/71 and
https://github.com/chaincodelabs/libmultiprocess/issues/122

The bug was caused by not having CustomBuildField and CustomReadField paired
together, so a lower priority CustomReadField was chosen that was not
compatible.