build: work around CMake quirk when building as a subdir of Bitcoin Core
When building as a subdir of Core and using depends, capnp includes may end up
being installed in a path like:
/dev/bitcoin/depends/HOST/include

It then detects that the install interface has been given an absolute path
inside of the project dir:
/dev/bitcoin/

In that case it throws an error because it wants paths within the project to
always be relative.

This shouldn't matter *at all* because we're never actually going to install
the lib anyway, but CMake populates the install interface despite the
EXCLUDE_FROM_ALL property being set.

Since other projects may wish to include libmultiprocess as a subdir and also
install it, don't use the subdir condition to check for this case. Instead,
test for the EXCLUDE_FROM_ALL on the directory, as if that's set, the parent
project won't be installing.